+++ anaconda/anaconda.py
@@ -1,0 +1,6 @@
+class Anaconda:
+    def __init__(self, name, species):
+        self.name = name
+        self.species = species
+        self.date_added = date.today()
+        self.slithering = False
