+++ animals.py
@@ -1,133 +1,21 @@
-from datetime import date
-
-
-class Llama:
-    def __init__(self, name, species):
-        self.name = name
-        self.species = species
-        self.date_added = date.today()
-        self.walking = False
-
-
-class Goat:
-    def __init__(self, name, species):
-        self.name = name
-        self.species = species
-        self.date_added = date.today()
-        self.walking = False
-
-
-class Duck:
-    def __init__(self, name, species):
-        self.name = name
-        self.species = species
-        self.date_added = date.today()
-        self.swimming = False
-
-
-class Turtle:
-    def __init__(self, name, species):
-        self.name = name
-        self.species = species
-        self.date_added = date.today()
-        self.swimming = False
-
-
-class Horse:
-    def __init__(self, name, species):
-        self.name = name
-        self.species = species
-        self.date_added = date.today()
-        self.walking = False
-
-
-class Cow:
-    def __init__(self, name, species):
-        self.name = name
-        self.species = species
-        self.date_added = date.today()
-        self.walking = False
-
-
-class Chicken:
-    def __init__(self, name, species):
-        self.name = name
-        self.species = species
-        self.date_added = date.today()
-        self.walking = False
-
-
-class CopperHead:
-    def __init__(self, name, species):
-        self.name = name
-        self.species = species
-        self.date_added = date.today()
-        self.slithering = False
-
-
-class RattleSnake:
-    def __init__(self, name, species):
-        self.name = name
-        self.species = species
-        self.date_added = date.today()
-        self.slithering = False
-
-
-class GardenSnake:
-    def __init__(self, name, species):
-        self.name = name
-        self.species = species
-        self.date_added = date.today()
-        self.slithering = False
-
-
-class Anaconda:
-    def __init__(self, name, species):
-        self.name = name
-        self.species = species
-        self.date_added = date.today()
-        self.slithering = False
-
-
-class CoralSnake:
-    def __init__(self, name, species):
-        self.name = name
-        self.species = species
-        self.date_added = date.today()
-        self.slithering = False
-
-
-class Goose:
-    def __init__(self, name, species):
-        self.name = name
-        self.species = species
-        self.date_added = date.today()
-        self.swimming = False
-
-
-class Swan:
-    def __init__(self, name, species):
-        self.name = name
-        self.species = species
-        self.date_added = date.today()
-        self.swimming = False
-
-
-class Bass:
-    def __init__(self, name, species):
-        self.name = name
-        self.species = species
-        self.date_added = date.today()
-        self.swimming = True
-
-    def __str__(self):
-        return f"My name is {self.name} and I am a {self.species}."
+from bass import Bass
+from chicken import Chicken
+from copperhead import CopperHead
+from coralsnake import CoralSnake
+from cow import Cow
+from duck import Duck
+from gardensnake import GardenSnake
+from goat import Goat
+from goose import Goose
+from horse import Horse
+from llama import Llama
+from rattlesnake import RattleSnake
+from swan import Swan
+from turtle import Turtle
 
 
 def main():
     myBass = Bass("Wall Singer", "Black Bass")
-    # myBass.name = "Wall Singer"
-    # myBass.species = "Black Bass"
 
     print(myBass)
