+++ chicken/__init__.py
@@ -1,0 +1,1 @@
+from .chicken import Chicken
