+++ coralsnake/__init__.py
@@ -1,0 +1,1 @@
+from .coralsnake import CoralSnake
