+++ cow/__init__.py
@@ -1,0 +1,1 @@
+from .cow import Cow
