+++ duck/__init__.py
@@ -1,0 +1,1 @@
+from .duck import Duck
