+++ turtle/turtle.py
@@ -1,0 +1,6 @@
+class Turtle:
+    def __init__(self, name, species):
+        self.name = name
+        self.species = species
+        self.date_added = date.today()
+        self.swimming = False
